apply_top_k_top_p_with_sorted算子中gm的偏移改为int64_t



# message auto-generated for no-merge-commit merge:
!263 merge apply_top_k_top_p_with_sorted into master

apply_top_k_top_p_with_sorted算子中gm的偏移改为int64_t

Created-by: wangxinwei21
Commit-by: wangxinwei328
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
apply_top_k_top_p_with_sorted算子中gm的偏移改为int64_t
## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->
[Bug-Report|缺陷反馈]: index\apply_top_k_top_p_with_sorted\op_kernel\apply_top_k_top_p_with_sorted.h batchOffset_变量_不符合“涉及gm内存偏移或大小必须使用int64表示”的编码风格，建议修改
## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->
![image.png](https://raw.gitcode.com/user-images/assets/7665709/72f7b336-c02a-4fe5-8b47-c7d362fec8c7/image.png 'image.png')
## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->
无
## 类型标签
<!--  [x] 表示选中 -->
- [x] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-nn!263

+++ index/apply_top_k_top_p_with_sorted/op_host/apply_top_k_top_p_with_sorted_tiling.h
@@ -26,19 +26,19 @@
 namespace optiling {
 
 BEGIN_TILING_DATA_DEF(ApplyTopKTopPWithSortedTilingData)
-    TILING_DATA_FIELD_DEF(uint32_t, batchSize);
-    TILING_DATA_FIELD_DEF(uint32_t, vocabSize);
-    TILING_DATA_FIELD_DEF(uint32_t, batchPerCore);
-    TILING_DATA_FIELD_DEF(uint32_t, tailBatch);
-    TILING_DATA_FIELD_DEF(uint32_t, blockNum);
-    TILING_DATA_FIELD_DEF(uint32_t, dataNumInit);
-    TILING_DATA_FIELD_DEF(uint32_t, dataNumInitAligned);
-    TILING_DATA_FIELD_DEF(uint32_t, ubFactorElement);
-    TILING_DATA_FIELD_DEF(uint32_t, ubFactorElementAligned);
-    TILING_DATA_FIELD_DEF(uint32_t, tailUbFactorElement);
-    TILING_DATA_FIELD_DEF(uint32_t, tailUbFactorElementAligned);
-    TILING_DATA_FIELD_DEF(uint32_t, calUbSize);
-    TILING_DATA_FIELD_DEF(uint32_t, iterateTimes);
+    TILING_DATA_FIELD_DEF(uint64_t, batchSize);
+    TILING_DATA_FIELD_DEF(uint64_t, vocabSize);
+    TILING_DATA_FIELD_DEF(uint64_t, batchPerCore);
+    TILING_DATA_FIELD_DEF(uint64_t, tailBatch);
+    TILING_DATA_FIELD_DEF(uint64_t, blockNum);
+    TILING_DATA_FIELD_DEF(uint64_t, dataNumInit);
+    TILING_DATA_FIELD_DEF(uint64_t, dataNumInitAligned);
+    TILING_DATA_FIELD_DEF(uint64_t, ubFactorElement);
+    TILING_DATA_FIELD_DEF(uint64_t, ubFactorElementAligned);
+    TILING_DATA_FIELD_DEF(uint64_t, tailUbFactorElement);
+    TILING_DATA_FIELD_DEF(uint64_t, tailUbFactorElementAligned);
+    TILING_DATA_FIELD_DEF(uint64_t, calUbSize);
+    TILING_DATA_FIELD_DEF(uint64_t, iterateTimes);
 END_TILING_DATA_DEF;
 REGISTER_TILING_DATA_CLASS(ApplyTopKTopPWithSorted, ApplyTopKTopPWithSortedTilingData)
